refactor code

+++ src/main/java/com/example/ecommerce/controller/ProductController.java
@@ -7,59 +7,71 @@ import com.example.ecommerce.model.Category;
 import com.example.ecommerce.model.User;
 import com.example.ecommerce.repository.CategoryRepository;
 import com.example.ecommerce.service.AuthenticationService;
+import com.example.ecommerce.service.CategoryService;
 import com.example.ecommerce.service.ProductService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("api/v1/product")
+@RequestMapping("api/v1/product/")
 public class ProductController {
 
     private final ProductService productService;
-    private final CategoryRepository categoryRepository;
+    private final CategoryService categoryService;
     private final AuthenticationService authenticationService;
-    public ProductController(ProductService productService, CategoryRepository categoryRepository,AuthenticationService authenticationService) {
+    public ProductController(ProductService productService, CategoryService categoryService,AuthenticationService authenticationService) {
         this.productService = productService;
-        this.categoryRepository = categoryRepository;
+        this.categoryService = categoryService;
         this.authenticationService = authenticationService;
     }
 
     @GetMapping(path = "all")
     public ResponseEntity<List<ProductDto>> getProducts() {
         List<ProductDto> products = productService.getAllProducts();
-        return new ResponseEntity<List<ProductDto>>(products, HttpStatus.OK);
+        return new ResponseEntity<>(products, HttpStatus.OK);
     }
 
-    @PostMapping
-    public ResponseEntity<ApiResponse> createProduct(@RequestParam("token") String token, @RequestBody ProductDto productDto) {
+    @PostMapping(path = "create")
+    public ResponseEntity<ApiResponse> createProduct(@RequestParam("token") String token, @Valid @RequestBody ProductDto productDto) {
         authenticationService.authenticate(token);
         if(!authenticationService.isAuthorized(token)){
             return new ResponseEntity<>(new ApiResponse(false, "access denied"), HttpStatus.FORBIDDEN);
         }
-        Optional<Category> optionalCategory = categoryRepository.findById(productDto.getCategoryId());
+        Optional<Category> optionalCategory = categoryService.findCategoryById(productDto.getCategoryId());
         if(!optionalCategory.isPresent()) {
-            return new ResponseEntity<>(new ApiResponse(false, "category does not exists"), HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new ApiResponse(false, "category does not exists"), HttpStatus.CONFLICT);
         }
-        productService.createProduct(productDto, optionalCategory.get());
+        Category category = optionalCategory.get();
+        productService.createProduct(productDto, category);
         return new ResponseEntity<>(new ApiResponse(true, "product has been created"),HttpStatus.CREATED);
     }
 
-    @PutMapping(path = "{productId}")
-    public ResponseEntity<ApiResponse> updateProduct(@PathVariable("productId") Long productId,@RequestBody ProductDto productDto) throws Exception {
-        Optional<Category>optionalCategory = categoryRepository.findById(productDto.getCategoryId());
-        if(!optionalCategory.isPresent()){
-            return new ResponseEntity<ApiResponse>(new ApiResponse(false, "category does not exists"), HttpStatus.BAD_REQUEST);
+    @PutMapping(path = "update/{productId}")
+    public ResponseEntity<ApiResponse> updateProduct(@RequestParam("token") String token, @PathVariable("productId") Long productId, @Valid @RequestBody ProductDto productDto) throws Exception {
+        authenticationService.authenticate(token);
+        if(!authenticationService.isAuthorized(token)){
+            return new ResponseEntity<>(new ApiResponse(false, "access denied"), HttpStatus.FORBIDDEN);
         }
-        productService.updateProduct(productDto, productId);
+        Optional<Category>optionalCategory = categoryService.findCategoryById(productDto.getCategoryId());
+        if(!optionalCategory.isPresent()){
+            return new ResponseEntity<ApiResponse>(new ApiResponse(false, "category does not exists"), HttpStatus.CONFLICT);
+        }
+        Category category = optionalCategory.get();
+        productService.updateProduct(productId, productDto, category);
         return new ResponseEntity<>(new ApiResponse(true, "product has been updated"), HttpStatus.OK);
     }
 
     @DeleteMapping(path = "{productId}")
-    public ResponseEntity<ApiResponse> deleteProduct(@PathVariable("productId") Long productId) throws Exception {
+    public ResponseEntity<ApiResponse> deleteProduct(@RequestParam("token") String token, @PathVariable("productId") Long productId) throws Exception {
+        authenticationService.authenticate(token);
+        if(!authenticationService.isAuthorized(token)){
+            return new ResponseEntity<>(new ApiResponse(false, "access denied"), HttpStatus.FORBIDDEN);
+        }
         productService.deleteProduct(productId);
         return new ResponseEntity<>(new ApiResponse(true, "product has been deleted"), HttpStatus.OK);
     }
